Group similar API calls
