+++ plugins/polio/js/src/domains/Budget/hooks/api/useGetBudgetProcessAvailableRounds.ts
@@ -36,3 +36,30 @@ export const useAvailableRoundsDependentDropdowns = (): UseQueryResult<
         },
     });
 };
+
+export const useGetAvailableRoundsForBudgetProcess = (
+    campaignID: string,
+    budgetProcessId: number,
+): UseQueryResult<OptionsRounds[], Error> => {
+    const select = useCallback((data: OptionsRounds[]) => {
+        return data.map((round: OptionsRounds) => {
+            return {
+                value: round.value,
+                label: formatRoundNumber(round.label),
+                campaign_id: round.campaign_id,
+            };
+        });
+    }, []);
+
+    return useSnackQuery({
+        queryKey: ['new_budget_process_dropdowns'],
+        queryFn: () =>
+            getRequest(
+                `/api/polio/budget/available_rounds/?campaign_id=${campaignID}&budget_process_id=${budgetProcessId}`,
+            ),
+        options: {
+            keepPreviousData: true,
+            select,
+        },
+    });
+};
